backend_django/requirements.txt
Django
psycopg2-binary
djangorestframework
django-phonenumber-field[phonenumbers]
python-dotenv
djangorestframework-simplejwt